file mapping done. need to decrypt the files. no compilation errors. one conversion warning, i dont know what to do about it :(

+++ MapleLib/src/Wz/WzBinaryReader.h
@@ -23,7 +23,7 @@ namespace Wz
 		std::string ReadString(long offset) { return ReadString(offset, false); }
 		std::string ReadString(long offset, bool readByte)
 		{
-			long CurrentOffset = m_FileStream.tellg();
+			long CurrentOffset = (long)m_FileStream.tellg();
 			m_FileStream.seekg(offset);
 			if (readByte)
 			{
